Non-parametric tests for the two-sample problem based on order statistics and power comparisons

tnl.test performs a nonparametric test for two sample test on vectors of data.

ptnl gives the distribution function of Tn(ℓ) against the specified quantiles.

dtnl gives the density of Tn(ℓ) against the specified quantiles.

qtnl gives the quantile function of Tn(ℓ) against the specified probabilities.

rtnl generates random values from Tn(ℓ).

tnl_mean gives an expression for E(Tn(ℓ)) under H0:F=G.

ptnl.lehmann gives the distribution function of Tn(ℓ) under Lehmann alternatives.

dtnl.lehmann gives the density of Tn(ℓ) under Lehmann alternatives.

qtnl.lehmann gives the quantile function of Tn(ℓ) against the specified probabilities under Lehmann alternatives.

rtnl.lehmann generates random values from Tn(ℓ) under Lehmann alternatives.

tnl.test(x, y, l, exact = "NULL")

ptnl(q, n, m, l, exact = "NULL", trial = 1e+05)

dtnl(k, n, m, l, exact = "NULL", trial = 1e+05)

qtnl(p, n, m, l, exact = "NULL", trial = 1e+05)

rtnl(N, n, m, l)

tnl_mean(n., m., l)

ptnl.lehmann(q, n., m., l, gamma)

dtnl.lehmann(k, n., m., l, gamma)

qtnl.lehmann(p, n., m., l, gamma)

rtnl.lehmann(N, n., m., l, gamma)

Arguments

x

the first (non-empty) numeric vector of data values.

y

the second (non-empty) numeric vector of data values.

l

class parameter of Tn(ℓ).

exact

the method that will be used. "NULL" or a logical indicating whether an exact should be computed. See 'Details' for the meaning of NULL.

n, m

samples size.

trial

number of trials for simulation.

k, q

vector of quantiles.

p

vector of probabilities.

N

number of observations. If length(N) > 1, the length is taken to be the number required.

n., m.

samples size.

gamma

parameter of Lehmann alternative.

Value

tnl.test returns a list with the following components

statistic:

the value of the test statistic.

p.value:

the p-value of the test.

ptnl returns a list with the following components

method:

The method that was used (exact or simulation). See 'Details'.

cdf:

distribution function of Tn(ℓ) against the specified quantiles.

dtnl returns a list with the following components

method:

The method that was used (exact or simulation). See 'Details'.

pmf:

density of Tn(ℓ) against the specified quantiles.

qtnl returns a list with the following components

method:

The method that was used (exact or simulation). See 'Details'.

quantile:

quantile function against the specified probabilities.

rtnl return N of the generated random values.

tnl_mean return the mean of Tn(ℓ).

ptnl.lehmann return vector of the distribution under Lehmann alternatives against the specified gamma.

dtnl.lehmann return vector of the density under Lehmann alternatives against the specified gamma.

qtnl.lehmann returns a quantile function against the specified probabilities under Lehmann alternatives.

rtnl.lehmann return N of the generated random values under Lehmann alternatives.

Details

A non-parametric two-sample test is performed for testing null hypothesis H0:F=G against the alternative hypothesis H1:F ≠ G. The assumptions of the Tn(ℓ) test are that both samples should come from a continuous distribution and the samples should have the same sample size.
Missing values are silently omitted from x and y.
Exact and simulated p-values are available for the Tn(ℓ) test. If exact ="NULL" (the default) the p-value is computed based on exact distribution when the sample size is less than 11. Otherwise, p-value is computed based on a Monte Carlo simulation. If exact ="TRUE", an exact p-value is computed. If exact="FALSE" , a Monte Carlo simulation is performed to compute the p-value. It is recommended to calculate the p-value by a Monte Carlo simulation (use exact="FALSE"), as it takes too long to calculate the exact p-value when the sample size is greater than 10.
The probability mass function (pmf), cumulative density function (cdf) and quantile function of Tn(ℓ) are also available in this package, and the above-mentioned conditions about exact ="NULL", exact ="TRUE" and exact="FALSE" is also valid for these functions.
Exact distribution of Tn(ℓ) test is also computed under Lehman alternative.
Random number generator of Tn(ℓ) test statistic are provided under null hypothesis in the library.

References

Karakaya K. et al. (2021). A Class of Non-parametric Tests for the Two-Sample Problem based on Order Statistics and Power Comparisons . Submitted paper.
Aliev F. et al. (2021). A Nonparametric Test for the Two-Sample Problem based on Order Statistics. Submitted paper.

Examples